[refractor] - context and components
