+++ src/components/cart/cartSubComponents/index.js
@@ -1,0 +1,2 @@
+export { CartDetails } from "./CartDetails";
+export { CartPriceDetails } from "./CartPriceDetails";
