+++ src/context/index.js
@@ -1,2 +1,2 @@
-export { StateProvider, useStateContext } from "./state-context";
+export { DataProvider, useData } from "./data-context";
 export { AuthProvider, useAuth } from "./auth-context";
